LPS-67663 SF - Rename packages and classes to avoid confusion.
