--- modules/apps/events-calendar/events-calendar-web/src/main/java/com/liferay/calendar/web/internal/upgrade/v1_0_2/UpgradePortletId.java
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.calendar.web.internal.upgrade.v1_0_2;
-
-import com.liferay.portal.kernel.upgrade.BaseUpgradePortletId;
-
-/**
- * @author Adam Brandizzi
- */
-public class UpgradePortletId extends BaseUpgradePortletId {
-
-	@Override
-	protected String[] getUninstanceablePortletIds() {
-		return new String[] {
-			"com_liferay_calendar_web_portlet_CalendarPortlet"
-		};
-	}
-
-}
