+++ modules/apps/events-calendar/events-calendar-web/src/main/java/com/liferay/events/calendar/web/internal/portlet/route/EventsCalendarFriendlyURLMapper.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.events.calendar.web.internal.portlet.route;
+
+import com.liferay.calendar.constants.CalendarPortletKeys;
+import com.liferay.portal.kernel.portlet.DefaultFriendlyURLMapper;
+import com.liferay.portal.kernel.portlet.FriendlyURLMapper;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Marcellus Tavares
+ */
+@Component(
+	immediate = true,
+	property = {
+		"com.liferay.portlet.friendly-url-routes=META-INF/friendly-url-routes/routes.xml",
+		"javax.portlet.name=" + CalendarPortletKeys.CALENDAR,
+		"service.ranking=1"
+	},
+	service = FriendlyURLMapper.class
+)
+public class EventsCalendarFriendlyURLMapper extends DefaultFriendlyURLMapper {
+}
